fastapi/requirements.txt
alembic==1.17.1
annotated-doc==0.0.3
antlr4-python3-runtime==4.9.3
anyio==4.11.0
Authlib==1.6.5
backcall==0.2.0
bcrypt==5.0.0
cachetools==6.2.1
capstone==6.0.0a5
colorama==0.4.6
colored-traceback==0.4.2
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.14
docopt==0.6.2
docstring_parser==0.17.0
fastapi==0.121.0
ffmpeg
ffmpeg-python==0.2.0
filelock==3.17.0
flatbuffers==25.2.10
fonttools==4.60.0
fsspec==2025.2.0
google-genai==1.49.0
graphviz==0.21
httpx-sse==0.4.3
huggingface-hub==0.29.3
humanfriendly==10.0
intervaltree==3.1.0
invoke==2.2.1
ipython==8.12.3
Jinja2==3.1.5
joblib==1.5.2
jsonref==1.1.0
kiwisolver==1.4.9
llvmlite==0.44.0
Mako==1.3.10
matplotlib==3.10.6
mcp==1.20.0
more-itertools==10.7.0
mpmath==1.3.0
numpy==2.2.2
omegaconf==2.3.0
openai-whisper==20250625
opencv-python==4.11.0.86
paramiko==4.0.0
pickleshare==0.7.5
pillow==11.1.0
pipreqs==0.5.0
polib==1.2.0
protobuf
pwntools==4.15.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyautogen==0.10.0
pydantic-settings==2.11.0
pyelftools==0.32
Pygments==2.19.1
PyJWT==2.10.1
PyNaCl==1.6.0
pyparsing==3.2.5
pyserial==3.5
python-multipart==0.0.20
pytz==2025.1
pyzmq==26.4.0
rapidocr==2.1.0
regex==2024.11.6
ROPGadget==7.7
rsa==4.9.1
ruamel.yaml.clib==0.2.14
rubicon-objc==0.5.1
safetensors==0.5.3
scikit-learn==1.7.2
scipy==1.16.2
sentencepiece==0.2.1
setuptools==69.5.1
socksio==1.0.0
sortedcontainers==2.4.0
speechbrain==1.0.3
sqlalchemy-spanner==1.17.1
sqlparse==0.5.3
srt==3.5.3
sse-starlette==3.0.3
starlette==0.49.3
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
transformers==4.49.0
tzdata==2025.1
tzlocal==5.3.1
unicorn==2.1.1
unix-ar==0.2.1
uvicorn==0.38.0
watchdog==6.0.0
websockets==15.0.1
wheel==0.43.0
yarg==0.1.9
